BackEnd/Database/IngestionScripts/requirements.txt
cffi==1.16.0
cryptography==41.0.5
oracledb==1.4.2
pycparser==2.21
python-dotenv==0.19.0
pandas
